+++ client/src/constants.ts
@@ -2,5 +2,7 @@ export enum Configuration {
     BaseSection = "twineTweeLanguage",
     FilesInclude = "files.include",
     FilesExclude = "files.exclude",
+    DownloadStoryFormat = "build.downloadStoryFormat",
+    StoryFormatDirectory = "build.storyFormatDirectory",
     Twee3WarningUnknownMacro = "twee-3.warning.unknownMacro",
 }